Speed up `one_of?`
Speed up `one_of?` by defining it as `false`, and when `one_of` is used,
redefining the method as `true`.

In profiles, I noticed that `oneOf?` was taking _some_ time. Looping
through the directive list each time feels wasteful.

This approach borrows from Rails' `Class#class_attribute` and redefines
the `one_of?` when calling the `one_of` class method. This respects the
fact that directives are inherited from parent classes; so is the
`one_of?` class method. That said, this is not 100% backward compatible
because users could be adding the OneOf directive without using the
`one_of` DSL, but... it still feels pretty safe?